backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
huggingface-hub==0.19.4
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.23
alembic==1.12.1
aiofiles==23.2.1
tenacity==8.2.3 